Refactor html tree parsing
